Renamed netcdf_groundwater to threedi_results and NetcdfGroundwaterDataSource to ResultData
