+++ test/test_datasources.py
@@ -23,9 +23,9 @@ try:
 except ImportError:
     print("Can't import Spatialite.")
     Spatialite = None
-from datasource.netcdf_groundwater import find_h5_file, find_aggregation_netcdf
-from ThreeDiToolbox.datasource.netcdf_groundwater import (
-    NetcdfGroundwaterDataSource,
+from datasource.threedi_results import find_h5_file, find_aggregation_netcdf
+from ThreeDiToolbox.datasource.threedi_results import (
+    ResultData,
     find_aggregation_netcdf_gw,
 )
 from ThreeDiToolbox.test.utilities import TemporaryDirectory
@@ -128,24 +128,24 @@ class TestSpatialiteDataSource(unittest.TestCase):
 class TestNetcdfGroundwaterDataSource(unittest.TestCase):
     def test_constructor(self):
         """Test empty constructor."""
-        NetcdfGroundwaterDataSource()
+        ResultData()
 
     def test_sanity(self):
-        nds = NetcdfGroundwaterDataSource()
+        nds = ResultData()
         m = mock.MagicMock()
         nds._datasource = m
         # sanity test
         self.assertEqual(nds.ds, m)
 
     @mock.patch(
-        "ThreeDiToolbox.datasource.netcdf_groundwater.NetcdfGroundwaterDataSource.available_subgrid_map_vars",
+        "ThreeDiToolbox.datasource.netcdf_groundwater.ResultData.available_subgrid_map_vars",
         ["s1"],
     )
     @mock.patch(
-        "ThreeDiToolbox.datasource.netcdf_groundwater.NetcdfGroundwaterDataSource.gridadmin_result"
+        "ThreeDiToolbox.datasource.netcdf_groundwater.ResultData.gridadmin_result"
     )
     def test_get_timeseries(self, gridadmin_result_mock):
-        nds = NetcdfGroundwaterDataSource()
+        nds = ResultData()
         m = mock.MagicMock()
         nds._datasource = m
         nds.get_timeseries("s1", 3)
