+++ threedi_statistics/tools/statistics.py
@@ -71,7 +71,7 @@ class DataSourceAdapter(Proxy):
 
     @property
     def has_groundwater(self):
-        return self.obj.__class__.__name__ == "NetcdfGroundwaterDataSource"
+        return self.obj.__class__.__name__ == "ResultData"
 
     @property
     def timestamps(self):
